move dice functions to its own file
