+++ ArcParser.py
@@ -1,6 +1,5 @@
 # ArcPy
 import json
-import random
 
 import discord
 
@@ -8,105 +7,13 @@ import tatsu.exceptions
 from tatsu import parse
 from tatsu.util import asjson
 
-GRAMMAR = """
-@@grammar::CALC
-start = expression $ ;
-
-expression
-    =
-    | left:expression op:'+' ~ right:term
-    | left:expression op:'-' ~ right:term
-    | term
-    ;
-    
-term
-    =
-    | left:term op:'*' ~ right:dice
-    | left:term op:'/' ~ right:dice
-    | dice
-    ;
-
-dice
-    =
-    | left:dice op:/[d]+/ ~ right:factor
-    | factor
-    ;
-    
-factor
-    =
-    | '(' ~ @:expression ')'
-    | number
-    ;
-    
-number = /\d+/;
-"""
+import dice
 
 class Settings:
     def __init__(self) -> None:
         with open("settings.json") as f:
             j = json.load(f)
         self.prefix = j["prefix"]
-
-class DiceSemantics:
-    def number(self, ast):
-        return int(ast)
-
-def operate(var2: str, operator: str, var1: str, dList):
-    outputVar = 0.0
-    if operator == "+":
-        outputVar = float(var1)+float(var2)
-    elif operator == "-":
-        outputVar = float(var1)-float(var2)
-    elif operator == "*":
-        outputVar = float(var1)*float(var2)
-    elif operator == "/":
-        outputVar = float(var1)/float(var2)
-    elif operator == "d":
-        return dice_parse(int(var1), int(var2), dList)
-    if outputVar.is_integer(): return int(outputVar)
-    return outputVar
-
-def calculate(expression, dList):
-    expression = expression.replace('["(",', "")
-    expression = expression.replace(',")"]', "")
-    expression = expression.replace('"', "")
-    expression = expression.replace(",", "")
-
-    data = []
-    digitStack = []
-    for e in expression:
-        if e.isdigit():
-            digitStack.append(e)
-        elif e == "]":
-            if len(digitStack)>0:
-                num: str = ""
-                for f in digitStack:
-                    num = num + f
-                data.append(num)
-                digitStack = []
-            
-            new_data = operate(data.pop(), data.pop(), data.pop(), dList)
-            data.pop()
-            data.append(str(new_data))
-        else:
-            if len(digitStack)>0:
-                num: str = ""
-                for f in digitStack:
-                    num = num + f
-                data.append(num)
-                digitStack = []
-            data.append(e)
-    return data[0]
-
-def dice_parse(x:int, y:int, dList: list[list[int]]) -> int:
-    s = []
-    sum = 0
-    for _ in range(x):
-        currVal = random.randint(1, y)
-        s.append(currVal)
-        sum += currVal
-    dList.append(s)
-    return sum
     
 def main():
     with open('AUTHTOKEN.txt', 'r') as f:
@@ -130,6 +37,7 @@ def main():
         ### ping
         if message.content.startswith("ArcPing"):
             await message.channel.send("Hello world!")
+        ### COMMANDS
         ### dice parsing
         elif message.content.startswith(settings.prefix):
             expression = message.content.removeprefix(settings.prefix)
@@ -137,7 +45,7 @@ def main():
                 await message.channel.send(f"```\n{int(expression)}\n```")
             else:
                 try:
-                    ast = parse(GRAMMAR, expression, semantics = DiceSemantics())
+                    ast = parse(dice.GRAMMAR, expression, semantics = dice.DiceSemantics())
                 except tatsu.exceptions.ParseException as e:
                     await message.channel.send(f"```ParseException: {e}```")
                     return
@@ -145,7 +53,7 @@ def main():
                 await message.channel.send(ast)
                 dList: list[list[int]] = []
                 await message.channel.send(f"```json\n{j}\n```")
-                await message.channel.send(f'```# {calculate(j, dList)}\nDetails:({expression}) {dList}```')
+                await message.channel.send(f'```# {dice.calculate(j, dList)}\nDetails:({expression}) {dList}```')
 
     client.run(TOKEN)
